+++ tmsm-frontend-optimizations.php
@@ -16,7 +16,7 @@
  * Plugin Name:       TMSM Frontend Optimizations
  * Plugin URI:        https://github.com/thermesmarins/tmsm-frontend-optimizations
  * Description:       Frontend Optimizations for Thermes Marins de Saint-Malo
- * Version:           1.0.1
+ * Version:           1.0.2
  * Author:            Nicolas Mollet
  * Author URI:        https://github.com/nicomollet
  * License:           GPL-2.0+
@@ -35,7 +35,7 @@ if ( ! defined( 'WPINC' ) ) {
  * Start at version 1.0.0 and use SemVer - https://semver.org
  * Rename this for your plugin and update it as you release new versions.
  */
-define( 'TMSM_FRONTEND_OPTIMIZATIONS_VERSION', '1.0.1' );
+define( 'TMSM_FRONTEND_OPTIMIZATIONS_VERSION', '1.0.2' );
 
 /**
  * The code that runs during plugin activation.